requirements.txt
django-oscar
# Oscar extensions
django-oscar-stores==0.5.1
django-oscar-paypal==0.8

# We need PostGIS as the stores extension uses GeoDjango.
psycopg2==2.5.1

# To use Solr, we need pysolr
pysolr==3.1

lxml==3.2.4
cssselect==0.9.1
BeautifulSoup==3.2.1
raven==4.0.3

# extras
django-smart-selects
pillow
django-ajaximage
django-iban
